Create JB_Solaris.py
This a (still incomplete) adaptation for the John Bowen Solaris synthesizer. The adaptation is able to discover a Solaris unit, receive an Edit Buffer and find the name of the preset, but it's not able to send a preset so far, I do not know if the Solaris sysex implementation is able to accept it.